ML.methods {sem}	R Documentation

Methods for sem Objects Fit Using the objectiveML, objectiveGLS, objectiveFIML, msemObjectiveML, and msemObjectiveGLS Objective Functions

Description

These functions are for objects fit by sem using the objectiveML (multivariate-normal full-information maximum-likelihood), link{objectiveFIML} (multivariate-normal full-information maximum-likihood in the presence of missing data), objectiveGLS (generalized least squares), and msemObjectiveML (multigroup multivariate-normal FIML) objective functions.

Usage

## S3 method for class 'objectiveML'
anova(object, model.2, robust=FALSE, ...)
## S3 method for class 'objectiveFIML'
anova(object, model.2, ...)

## S3 method for class 'objectiveML'
logLik(object, ...)
## S3 method for class 'objectiveFIML'
logLik(object, saturated=FALSE, 
    intercept="Intercept", iterlim=1000, ...)
## S3 method for class 'objectiveML'
deviance(object, ...)
## S3 method for class 'objectiveFIML'
deviance(object, saturated.logLik, ...)
## S3 method for class 'msemObjectiveML'
deviance(object, ...) 
## S3 method for class 'objectiveML'
AIC(object, ..., k)
## S3 method for class 'objectiveFIML'
AIC(object, saturated.logLik, ..., k)
## S3 method for class 'msemObjectiveML'
AIC(object, ..., k)
## S3 method for class 'objectiveML'
AICc(object, ...)
## S3 method for class 'objectiveFIML'
AICc(object, saturated.logLik, ...)
## S3 method for class 'msemObjectiveML'
AICc(object, ...)
## S3 method for class 'objectiveML'
BIC(object, ...)
## S3 method for class 'objectiveFIML'
BIC(object, saturated.logLik, ...)
## S3 method for class 'msemObjectiveML'
BIC(object, ...)
## S3 method for class 'objectiveML'
CAIC(object, ...)
## S3 method for class 'objectiveFIML'
CAIC(object, saturated.logLik, ...)

## S3 method for class 'objectiveML'
print(x, ...)
## S3 method for class 'objectiveGLS'
print(x, ...)
## S3 method for class 'objectiveFIML'
print(x, saturated=FALSE, ...)
## S3 method for class 'msemObjectiveML'
print(x, ...)
## S3 method for class 'msemObjectiveGLS'
print(x, ...)

## S3 method for class 'objectiveML'
summary(object, digits=getOption("digits"), 
    conf.level=.90, robust=FALSE, analytic.se=object$t <= 500, 
    fit.indices=c("GFI", "AGFI", "RMSEA", "NFI", "NNFI", "CFI", "RNI",
      "IFI", "SRMR", "AIC", "AICc", "BIC", "CAIC"), ...)
## S3 method for class 'objectiveFIML'
summary(object, digits=getOption("digits"), conf.level=.90,
    fit.indices=c("AIC", "AICc", "BIC", "CAIC"),
    saturated=FALSE, intercept="Intercept", saturated.logLik, ...)
## S3 method for class 'objectiveGLS'
summary(object, digits=getOption("digits"), conf.level=.90, 
    fit.indices=c("GFI", "AGFI", "RMSEA", "NFI", "NNFI", "CFI", "RNI", "IFI", "SRMR"),
    ...)
## S3 method for class 'msemObjectiveML'
summary(object, digits=getOption("digits"), 
    conf.level=.90, robust=FALSE, 
    analytic.se=object$t <= 500,
    fit.indices=c("GFI", "AGFI", "RMSEA", "NFI", "NNFI", "CFI", "RNI", 
      "IFI", "SRMR", "AIC", "AICc", "BIC"), ...)
## S3 method for class 'msemObjectiveGLS'
summary(object, digits=getOption("digits"), 
    conf.level=.90,
    fit.indices=c("GFI", "AGFI", "RMSEA", "NFI", "NNFI", 
      "CFI", "RNI", "IFI", "SRMR"), ...)

Arguments

object, model.2, x
an object inheriting from class objectiveML, objectiveGLS, objectiveFIML, msemObjectiveML, or msemObjectiveGLS.

robust
if TRUE, compute robust standard errors or test.

fit.indices
a character vector of “fit indices” to report; the allowable values are those given in Usage above, and vary by the objective function. If the argument isn't given then the fit indices reported are taken from the R fit.indices option; if this option isn't set, then only the AIC and BIC are reported for models fit with objectiveML, objectiveFIML, or msemObjectiveML, and no fit indices are reported for models fit with objectiveGLS or msemObjectiveGLS.

k, ...
ignored.

digits
digits to be printed.

conf.level
level for confidence interval for the RMSEA index (default is .9).

analytic.se
use analytic (as opposed to numeric) coefficient standard errors; default is TRUE where analytic standard errors are available if there are no more than 100 parameters in the model and FALSE otherwise.

saturated
if TRUE (the default is FALSE); compute the log-likelihood (and statistics that depend on it) for the saturated model when the objective function is FIML in the presence of missing data. This can be computationally costly.

intercept
the name of the intercept regressor in the raw data, to be used in calculating the saturated log-likelihood for the FIML estimator; the default is "Intercept".

saturated.logLik
the log-likelihood for the saturated model, as returned by logLik with saturated=TRUE; if absent, this will be computed and the computation can be time-consuming.

iterlim
iteration limit used by the nlm optimizer to compute the saturated log-likelihood for the FIML estimator with missing data; defaults to 1000.

Author(s)

John Fox jfox@mcmaster.ca and Jarrett Byrnes

References

See sem.

See Also

sem, objective.functions, modIndices.objectiveML

[Package sem version 3.1-11 Index]