sem-deprecated {sem}	R Documentation

Deprecated Functions in the sem Package

Description

These functions are provided for compatibility with older versions of the sem package only, and may be removed eventually. Although an effort has been made to insure backwards-compatibility, commands that worked in versions of the sem package prior to version 2.0-0 will not necessarily work in version 2.0-0 and beyond, or may not work in the same manner.

Usage

boot.sem(...)
mod.indices(...)
normalized.residuals(...)
path.diagram(...)
raw.moments(...)
read.moments(...)
specify.model(...)
standardized.coefficients(...)
standardized.residuals(...)
std.coef(...)

Arguments

...
pass arguments down to replacements for deprecated functions.

Details

boot.sem is now a synonym for the bootSem function.

mod.indices is now a synonym for modIndices.

normalized.residuals is now a synonym for normalizedResiduals.

path.diagram is now a synonym for pathDiagram.

raw.moments is now a synonym for rawMoments.

read.moments is now a synonym for readMoments.

specify.model is now a synonym for specifyModel.

standardized.coefficients and std.coef are now synonyms for the standardizedCoefficients and stdCoef functions.

standardized.residuals is now a synonym for standardizedResiduals.

[Package sem version 3.1-11 Index]